test for binary compatibility
